pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "floorplan"
version = "0.1.0"
description = "Wi-Fi FTM/RTT indoor positioning and surveillance system"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "FLOORPLAN Team" }]

dependencies = [
    "click>=8.1",
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "websockets>=12.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "scapy>=2.5",
    "pyroute2>=0.7",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "rich>=13.0",
    "matplotlib>=3.8",
    "reportlab>=4.0",
    "aiosqlite>=0.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
]

[project.scripts]
floorplan = "floorplan.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
